common.h: ARRAY_SIZE() now returns size_t instead of int.
This fixes a lot of PVS-Studio V220 warnings due to casting from size_t
to int and back to size_t.

PVS-Studio V220 (Hi): Suspicious sequence of types castings: memsize ->
	32-bit integer -> memsize.

New ARRAY_SIZE_I() macro for cases where an int is definitely needed, e.g.
the various text functions and casts from enums.

Incidentally, eliminating the V220 warnings reduced the code size of
rpcli by a bit, though it increased TextFuncsTest:

Code size differences: (64-bit Gentoo Linux, gcc-11.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
1467693   92184   30056 1589933  1842ad rpcli
1467549   92184   30056 1589789  18421d rpcli
   -144       0       0    -144     -90 Difference

   text    data     bss     dec     hex filename
 468730   11368    1504  481602   75942 TextFuncsTest
 468814   11368    1504  481686   75996 TextFuncsTest
    +84       0       0     +84     +54 Difference

+++ src/gtk/OptionsMenuButton.cpp
@@ -706,7 +706,7 @@ options_menu_button_update_op(OptionsMenuButton *widget,
 	GList *l = gtk_container_get_children(GTK_CONTAINER(widget->menuOptions));
 
 	// Skip the standard actions and separator.
-	for (int i = 0; i < ARRAY_SIZE(stdacts) && l != nullptr; i++) {
+	for (size_t i = 0; i < ARRAY_SIZE(stdacts) && l != nullptr; i++) {
 		l = l->next;
 	}
